third_party/WebKit/LayoutTests/webmidi/permission-expected.txt

Test if sysex permission request is handled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testRunner.setMIDISysExPermission is defined.
PASS navigator.requestMIDIAccess is defined.
PASS promise = navigator.requestMIDIAccess({sysex: true}) did not throw exception.
PASS promise is defined.
PASS promise.then is defined.
PASS sysex permission request is successfully rejected.
PASS promise = navigator.requestMIDIAccess({sysex: true}) did not throw exception.
PASS promise is defined.
PASS promise.then is defined.
PASS sysex permission request is successfully accepted.
PASS all permission request rests pass successfully.
PASS successfullyParsed is true

TEST COMPLETE
